requirements.txt
amqp==2.5.2
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
billiard==3.6.3.0
boto3==1.12.45
botocore==1.15.45
celery==4.4.2
certifi==2019.11.28
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.4
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-cron==0.5.1
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-import-export==2.1.0
django-log-request-id==1.4.1
django-model-utils==4.0.0
django-mysql==3.5.0
django-redis==4.11.0
django-request-logging==0.7.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-writable-nested==0.6.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.0
importlib-resources==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
jmespath==0.9.5
kombu==4.6.8
MarkupPy==1.14
mongoengine==0.20.0
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
parso==0.6.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.1.1
pipenv==2018.11.26
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.5.2
requests==2.9.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.10
virtualenv-clone==0.5.3
wcwidth==0.1.8
xlrd==1.2.0
xlwt==1.3.0
xmljson==0.2.1
zipp==3.1.0
